+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### next [????.??.??]
+* Depend on `quickcheck-instances-0.3.17` or later.
+
 ### 3.7.1 [2017.12.29]
 * Don't define a `TextShow GCStats` instance on GHC 8.4 or later, as `GCStats`
   has been removed from `base`. This fixes the build on GHC 8.4.